feat(pattern): modularize pattern_generation
