+++ src/dmd_toolbox/pattern/__init__.py
@@ -1,0 +1,6 @@
+from ._core import analogDMDPattern, binaryMask
+
+__all__ = [
+    "binaryMask",
+    "analogDMDPattern"
+]
