--- src/dmd_toolbox/pattern_generation/__init__.py
@@ -1,7 +1,0 @@
-from .utils import analogBWImage, gaussian, quadratic
-
-__all__ = [
-    "analogBWImage",
-    "gaussian",
-    "quadratic"
-]
